+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Allow custom parser
 - Marketplace.id and Marketplace.ids shorthands
 
 ## [3.0.0] - 2024-10-04